% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attack.R
\name{attack}
\alias{attack}
\title{attack}
\usage{
attack(entity, attack)
}
\arguments{
\item{entity}{a character entity}

\item{attack}{a attack option}
}
\value{
Returns a damage value corresponding to the attack
}
\description{
This function retrieves the damage values from the character.json file for a corresponding attack.
}
\examples{
attack("hero","punch")

}
