% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_scene.R
\name{create_scene}
\alias{create_scene}
\title{create_scene}
\usage{
create_scene(foe = NULL, hero = NULL, menu = NULL, f.hp = NULL, h.hp = NULL)
}
\arguments{
\item{foe}{enemy opponent}

\item{hero}{player entity}

\item{menu}{menu type}

\item{f.hp}{opponent health value (max value 10)}

\item{h.hp}{player health value (max value 10)}
}
\value{
Generates a plot
}
\description{
This function creates the Scifighter battle scene
}
\examples{
\dontrun{
create_scene(foe="publisher", hero="hero", menu="menu_top_level", f.hp=10, h.hp=10)
}

}
