% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/encounter.R
\name{encounter}
\alias{encounter}
\title{encounter}
\usage{
encounter(foe, hero)
}
\arguments{
\item{foe}{opponent entity}

\item{hero}{player entity}
}
\value{
This function returns the match results
}
\description{
This functions initiates the encounter
}
\examples{
\dontrun{
encounter(foe="funder", hero="hero")
}

}
