% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/end_scene.R
\name{end_scene}
\alias{end_scene}
\title{end_scene}
\usage{
end_scene(entity = NULL)
}
\arguments{
\item{entity}{winner of the match}
}
\value{
Displays a plot
}
\description{
This function generates the post-battle end scene
}
\examples{
\dontrun{
end_scene("hero")
}

}
