% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamestart.R
\name{gamestart}
\alias{gamestart}
\title{gamestart}
\usage{
gamestart(nosound = NULL)
}
\arguments{
\item{nosound}{optional, in case of sound issue set this to true.}
}
\value{
Nothing
}
\description{
This function is required to start the game.
}
\examples{
\dontrun{
gamestart()
}

}
