% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_stats.R
\name{get_stats}
\alias{get_stats}
\title{get_stats}
\usage{
get_stats(x)
}
\arguments{
\item{x}{entity name}
}
\value{
Returns all stats of the corresponding entity.
}
\description{
This function retrieves the character stats corresponding to the entity.
}
\examples{
get_stats("phd.candidate")

}
