% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/encounter.R
\name{encounter}
\alias{encounter}
\title{encounter}
\usage{
encounter(foe, hero, nosound = NULL)
}
\arguments{
\item{foe}{opponent entity}

\item{hero}{player entity}

\item{nosound}{optional, Set this to TRUE to disable sound.}
}
\value{
This function returns the match results
}
\description{
This functions initiates the encounter
}
\examples{
\dontrun{
encounter(foe="funder", hero="hero")
}

}
